fix all-but-dask tests (#4511)
* fix all-but-dask tests

* whats new

* add dummy context
